Use legacy mapping of tiled op to processors for conv/pool.
The change to remove control overhead by using a single if-statement
instead of loops while mapping a tiled linalg op to workitems within a
workgroup seems to lead to correctness issue. The change is valid if
the number of iterations of the loop is less than or equal to the
workgroup size, which doesnt seem to be the case. More investigation
is needed.
For now, falling back to the loop method for convolution/pooling.

The change also enables the option to split the padding into a
separate option when needed. This is done to make it easier to
pin-point the issue, so that the complications from padding are
removed.

PiperOrigin-RevId: 318208338